+++ src/lib/components/Stepper/Step/Step.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export interface IStepProps {
+}
+
+export const Step = (props: IStepProps) => {
+	return (
+		<div>
+			coming soon step
+		</div>
+	);
+}