+++ src/lib/components/Stepper/StepLabel/StepLabel.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export interface IStepLabel {
+
+}
+
+export const StepLabel = (props: IStepLabel) => {
+	return (
+		<div>
+			coming soon StepLabel
+		</div>
+	)
+}