+++ src/lib/components/Stepper/index.ts
@@ -1,0 +1,3 @@
+export * from './Step';
+export * from './StepLabel';
+export * from './Stepper';