+++ public/changelog.md
@@ -1,3 +1,7 @@
+## unreleased 
+- stepper - added component
+
+
 ## 13.3.2
 - form - onEnter fix